13.7.File Uploads & Data Fixtures{Cleanup into a Private Method}
Cleanup into a Private Method

Though, I don't love having all of this logic right in the middle of this already-long function: it's not super obvious what it does. Let's do some cleanup: copy all of this. And at the bottom, create a new private function fakeUploadImage() that will return a string.
Paste all that logic and return the $this->uploaderHelper line. It selects a random image, uploads it and returns the path.
Back up top, delete all this stuff and say $imageFilename = $this->fakeUploadImage().

Let's run those fixtures one more time!
php bin/console doctrine:load:fixtures
When it finishes... we have some new files... and the homepage is shiny! That's a solid fixture system.

Next: we'll take our first step towards storing uploaded files in the cloud by integrating the gorgeous Flysystem library.