16.1.Flysystem: Streaming & Defensive Coding{Streaming}
There were a few minor problems with our new Flysystem integration. Let's clean them up before they bite us!

Streaming

The first is that using file_get_contents() eats memory: it reads the entire contents of the file into PHP's memory. That's not a huge deal for tiny files, but it could be a big deal if you start uploading bigger stuff. And, it's just not necessary.

For that reason, in general, when you use Flysystem, instead of using methods like ->write() or ->update(), you should use ->writeStream() or ->updateStream().

It works the same, except that we need to pass a stream instead of the contents. Create the stream with $stream = fopen($file->getPathname()) and, because we just need to read the file, use the r flag. Now, pass stream instead of the contents.

Yea... that's it! Same thing, but no memory issues. But we do need to add one more detail after: if is_resource($stream), then fclose($stream). The 'if' is needed because some Flysystem adapters close the stream by themselves.

Ok, for problem number two, go back to /admin/article. Log back in with password engage, edit an article, and go select an image - how about astronaut.jpg. Hit update and... it works! So what's the problem? Well, we just replaced an existing image with this new one. Does the old file still exist in our uploads directory? Absolutely! But it probably shouldn't. When an article image is updated, let's delete the old file.