16.2.Flysystem: Streaming & Defensive Coding{Deleting the Old File}
Deleting the Old File

In UploaderHelper, add a second argument - a nullable string argument called $existingFilename.

This is nullable because sometimes there may not be an existing file to delete. At the bottom, it's beautifully simple: if an $existingFilename was passed, then $this->filesystem->delete() and pass that the full path, which will be self::ARTICLE_IMAGE.'/'.$existingFilename.

In new(), you can really just pass null - there will not be an article image. But I'll pass getImageFilename() to be consistent.

Oh, and there's one other place we need update: ArticleFixtures. Down here, just pass null: we are never updating.

Try it! Here is the current astronaut image. Now, move over, upload rocket.jpg this time and update! Back in the directory... there's rocket and astronaut is gone! Love it!