17.4.Flysystem <3 LiipImagineBundle{The Resolver: Saving the Images to Flysystem}
The Resolver: Saving the Images to Flysystem

So, the 'data loader' is responsible for reading the original image. But, there's another important concept from LiipImagineBundle called 'resolvers'. Click down on the 'Flysystem Resolver' in their docs. The resolver is responsible for saving the thumbnail image back to the filesystem after all of the transformations. By default, no surprise, LiipImagineBundle writes things directly to the filesystem. So even if we moved Flysystem to s3, LiipImagineBundle would still be writing the thumbnail files back to our server - into the public/media directory.

 Tip

You can also completely offload the processing and storage of your files to a cloud service like rokka.io by leveraging LiipRokkaImagineBundle.